requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
python-multipart>=0.0.6
openai>=1.0.0
Pillow>=10.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0